feat: Add new components for About section including SectionTitle, FeatureDetail, UserCase, UserCases, WhyChooseUs, CallToAction, and PageHeader
